Extract shared GlyphPath utility from duplicated SVG and PDF glyph path code.
